fixed bugs and added features
Fixed bug with installing from Live USB created from a hybrid ISO. Added the ability to specify the size of the root and home partitions and added the ability to specify the filesystem type (ext4, xfs) for the root and home partitions. Added more commenting so the code is easier to understand.